chore(api): remove all Canva integration code
Canva is no longer used. This removes:
- canva_user_id profile field and getProfileByCanvaId() method
- canva feature flag from types, auth middleware, and mobile auth
- canvaUtils.ts (URL validation/processing)
- /user-templates/from-url Canva import endpoint
- Redis OAuthStateManager (Canva-only OAuth state)
- CSP entries for canva.com
- Canva OAuth columns from local_setup.sql
- Default template_type changed from 'canva' to 'template'
