+++ apps/api/database/postgres/local_setup.sql
@@ -26,14 +26,6 @@ CREATE TABLE IF NOT EXISTS profiles (
     presseabbinder TEXT,
     custom_antrag_gliederung TEXT,
     auth_source TEXT,
-    canva_access_token TEXT,
-    canva_refresh_token TEXT,
-    canva_token_expires_at TIMESTAMPTZ,
-    canva_user_id TEXT,
-    canva_display_name TEXT,
-    canva_email TEXT,
-    canva_scopes JSONB DEFAULT '[]',
-    canva_team_id TEXT,
     groups_enabled BOOLEAN DEFAULT FALSE,
     custom_generators BOOLEAN DEFAULT FALSE,
     database_access BOOLEAN DEFAULT FALSE,
