+++ apps/api/routes/auth/mobileAuth.ts
@@ -102,7 +102,6 @@ function sanitizeUserForResponse(user: UserProfile): Partial<UserProfile> {
     notebook,
     sharepic,
     anweisungen,
-    canva,
     labor_enabled,
     sites_enabled,
     chat,
@@ -132,7 +131,6 @@ function sanitizeUserForResponse(user: UserProfile): Partial<UserProfile> {
     notebook,
     sharepic,
     anweisungen,
-    canva,
     labor_enabled,
     sites_enabled,
     chat,
