+++ apps/api/types/express.d.ts
@@ -22,7 +22,6 @@ interface UserProfileShape {
   notebook: boolean;
   sharepic: boolean;
   anweisungen: boolean;
-  canva: boolean;
   labor_enabled: boolean;
   sites_enabled: boolean;
   chat: boolean;
@@ -32,7 +31,6 @@ interface UserProfileShape {
   video_editor: boolean;
   bundestag_api_enabled?: boolean;
   memory_enabled?: boolean;
-  canva_user_id?: string;
   created_at: Date | string;
   updated_at: Date | string;
   last_login?: Date | string;
