--- apps/api/utils/redis/OAuthStateManager.ts
@@ -1,283 +1,0 @@
-/**
- * Redis OAuth State Manager
- *
- * Manages OAuth state storage in Redis for reliable cross-session persistence.
- * Handles PKCE values, user context, and automatic expiration.
- */
-
-import * as redis from 'redis';
-import type { RedisClientType } from 'redis';
-import type { OAuthStateData, OAuthStats } from './types.js';
-import * as dotenv from 'dotenv';
-
-dotenv.config({ quiet: true });
-
-/**
- * Manages OAuth state in Redis with automatic expiration
- */
-class RedisOAuthStateManager {
-  private client: RedisClientType | null = null;
-  private isConnected: boolean = false;
-
-  constructor() {
-    this.init();
-  }
-
-  /**
-   * Initialize Redis connection
-   */
-  private async init(): Promise<void> {
-    try {
-      if (!process.env.REDIS_URL) {
-        console.warn(
-          '[Redis OAuth] No REDIS_URL configured - OAuth state will use fallback storage'
-        );
-        return;
-      }
-
-      console.log(
-        '[Redis OAuth] Initializing Redis client with URL:',
-        process.env.REDIS_URL.replace(/:\/\/(.*:)?(.*)@/, '://<user>:<password>@')
-      );
-
-      this.client = redis.createClient({
-        url: process.env.REDIS_URL,
-        socket: {
-          reconnectStrategy: (retries: number) => {
-            if (retries > 10) {
-              console.error('[Redis OAuth] Too many retry attempts');
-              return new Error('Too many retry attempts');
-            }
-            return Math.min(retries * 100, 3000);
-          },
-        },
-      });
-
-      this.client.on('error', (err: Error) => {
-        console.error('[Redis OAuth] Connection error:', err.message);
-        this.isConnected = false;
-      });
-
-      this.client.on('connect', () => {
-        console.log('[Redis OAuth] Connected to Redis server');
-        this.isConnected = true;
-      });
-
-      this.client.on('ready', () => {
-        console.log('[Redis OAuth] Redis client ready');
-        this.isConnected = true;
-      });
-
-      this.client.on('end', () => {
-        console.log('[Redis OAuth] Redis connection ended');
-        this.isConnected = false;
-      });
-
-      await this.client.connect();
-    } catch (error) {
-      console.error('[Redis OAuth] Failed to initialize Redis:', (error as Error).message);
-      this.isConnected = false;
-    }
-  }
-
-  /**
-   * Generate Redis key for OAuth state
-   */
-  private getStateKey(state: string): string {
-    return `oauth:canva:state:${state}`;
-  }
-
-  /**
-   * Store OAuth state in Redis
-   * @param state - OAuth state parameter
-   * @param data - OAuth data to store
-   * @param ttlSeconds - TTL in seconds (default: 600 = 10 minutes)
-   * @returns Success status
-   */
-  async storeState(
-    state: string,
-    data: OAuthStateData,
-    ttlSeconds: number = 600
-  ): Promise<boolean> {
-    try {
-      if (!this.isConnected || !this.client) {
-        console.warn('[Redis OAuth] Redis not available for storing state');
-        return false;
-      }
-
-      const key = this.getStateKey(state);
-      const stateData: OAuthStateData = {
-        ...data,
-        createdAt: Date.now(),
-        expiresAt: Date.now() + ttlSeconds * 1000,
-      };
-
-      await this.client.setEx(key, ttlSeconds, JSON.stringify(stateData));
-
-      console.log('[Redis OAuth] OAuth state stored', {
-        state: state.substring(0, 20) + '...',
-        userId: data.userId,
-        ttl: ttlSeconds,
-        expiresAt: new Date(stateData.expiresAt!).toISOString(),
-      });
-
-      return true;
-    } catch (error) {
-      console.error('[Redis OAuth] Error storing OAuth state:', (error as Error).message);
-      return false;
-    }
-  }
-
-  /**
-   * Retrieve and delete OAuth state from Redis
-   * @param state - OAuth state parameter
-   * @returns OAuth data or null if not found
-   */
-  async retrieveState(state: string): Promise<OAuthStateData | null> {
-    try {
-      if (!this.isConnected || !this.client) {
-        console.warn('[Redis OAuth] Redis not available for retrieving state');
-        return null;
-      }
-
-      const key = this.getStateKey(state);
-
-      // Get and delete in one atomic operation
-      const multi = this.client.multi();
-      multi.get(key);
-      multi.del(key);
-      const results = await multi.exec();
-
-      const dataString = results[0] as unknown as string | null;
-
-      if (!dataString) {
-        console.log('[Redis OAuth] OAuth state not found', {
-          state: state.substring(0, 20) + '...',
-        });
-        return null;
-      }
-
-      const data = JSON.parse(dataString) as OAuthStateData;
-
-      // Double-check expiration (Redis TTL should handle this, but be safe)
-      if (data.expiresAt && Date.now() > data.expiresAt) {
-        console.log('[Redis OAuth] OAuth state expired', {
-          state: state.substring(0, 20) + '...',
-          expiredAt: new Date(data.expiresAt).toISOString(),
-        });
-        return null;
-      }
-
-      console.log('[Redis OAuth] OAuth state retrieved', {
-        state: state.substring(0, 20) + '...',
-        userId: data.userId,
-        age: data.createdAt ? Math.round((Date.now() - data.createdAt) / 1000) + 's' : 'unknown',
-      });
-
-      return data;
-    } catch (error) {
-      console.error('[Redis OAuth] Error retrieving OAuth state:', (error as Error).message);
-      return null;
-    }
-  }
-
-  /**
-   * Check if Redis is available
-   */
-  isAvailable(): boolean {
-    return this.isConnected && this.client !== null;
-  }
-
-  /**
-   * Clean up expired OAuth states (manual cleanup - Redis TTL should handle most cases)
-   * @returns Number of states cleaned up
-   */
-  async cleanupExpiredStates(): Promise<number> {
-    try {
-      if (!this.isConnected || !this.client) {
-        return 0;
-      }
-
-      const pattern = 'oauth:canva:state:*';
-      const keys = await this.client.keys(pattern);
-
-      if (keys.length === 0) {
-        return 0;
-      }
-
-      let cleaned = 0;
-      const now = Date.now();
-
-      for (const key of keys) {
-        try {
-          const dataString = await this.client.get(key);
-          if (dataString && typeof dataString === 'string') {
-            const data = JSON.parse(dataString) as OAuthStateData;
-            if (data.expiresAt && now > data.expiresAt) {
-              await this.client.del(key);
-              cleaned++;
-            }
-          }
-        } catch (parseError) {
-          // If we can't parse the data, delete the key
-          await this.client.del(key);
-          cleaned++;
-        }
-      }
-
-      if (cleaned > 0) {
-        console.log(`[Redis OAuth] Manually cleaned up ${cleaned} expired OAuth states`);
-      }
-
-      return cleaned;
-    } catch (error) {
-      console.error('[Redis OAuth] Error during cleanup:', (error as Error).message);
-      return 0;
-    }
-  }
-
-  /**
-   * Get statistics about stored OAuth states
-   */
-  async getStats(): Promise<OAuthStats> {
-    try {
-      if (!this.isConnected || !this.client) {
-        return { available: false, count: 0 };
-      }
-
-      const pattern = 'oauth:canva:state:*';
-      const keys = await this.client.keys(pattern);
-
-      return {
-        available: true,
-        count: keys.length,
-        connected: this.isConnected,
-      };
-    } catch (error) {
-      console.error('[Redis OAuth] Error getting stats:', (error as Error).message);
-      return { available: false, count: 0, error: (error as Error).message };
-    }
-  }
-
-  /**
-   * Close Redis connection
-   */
-  async close(): Promise<void> {
-    try {
-      if (this.client) {
-        await this.client.quit();
-        console.log('[Redis OAuth] Redis connection closed');
-      }
-    } catch (error) {
-      console.error('[Redis OAuth] Error closing Redis connection:', (error as Error).message);
-    }
-  }
-}
-
-// Export singleton instance
-const redisOAuthStateManager = new RedisOAuthStateManager();
-
-export default redisOAuthStateManager;
-
-// Named export for modern imports
-export { redisOAuthStateManager, RedisOAuthStateManager };
